Add JSON Schemas for Driver and Layer Manifests
The JSON Schema format allows for automatic validation of a given manifest file.
This enhances the ability for layer and driver developers to check that their
manifest files are correct, as well as clarifying the exact format that the
loader expects manifest files to be in.

These files will be available inside the repo for reference, but the primary
location they will be made available from is:
https://schema.khornos.org/vulkan/

Each version of the Driver and Layer Manifest formats have their own schema,
allowing verification of whichever version the driver or layer intends to use.