Add Debug extension support to test layer
Debug utils, debug report, and debug marker are now supported directly in the
test layer, allowing easy testing of the behavior of layers which support
these debug extensions.

This allows testing the matrix of combinations between drivers, layers, and
loader support for these debug extensions.

This commit also adds a 'injected_extension' concept into the test_layer,
which allows recreating the behavior of layers which modify the extension
list returned by the driver.