Warn if apiVersion is below VK_API_VERSION_1_0 and above 0
Setting the apiVersion to an invalid value is an easy mistake to make.
But because it causes drivers to fail, the validation layers can not catch
it in time.

This commit makes it a 'fatal error' which forces log_message to write out
the message, regardless of whether or not VK_LOADER_DEBUG is set or if the
application has setup a debug callback in the pNext chain.