Remove -fno-strict-aliasing from builds
This was added in commit 72199501cb01497ddc0e429d561e2e7e17a841c4

It was purported to fix a hang, but was never fully verified if it actually
fixed the underlying issue (or if the issue came from buggy optimizing
compilers or drivers).

The chromium build does not enable -fno-strict-aliasing and hasn't had any
issues known to be caused by strict aliasing optimizations.

This commit also fixes debug report/debug utils code that was not properly
casing to uintptr_t before assinging or dereferencing function parameters.
The same logic occurs in wsi.h which as it assigned a pointer of VkIcdSurface
to the VkSurface parameter.