Change loader-unknown-chain into a STATIC library
loader-unknown-chain is now a STATIC library instead of an OBJECT
library. Object libraries have limitations compared to 'normal' libraries,
and this simplifies the CMake usage present.

This change includes a fix for building using Ninja, as there is a bug
in cmake related to static libraries that contain assembly code that
affect builds using MSVC and clang-cl.