Add unknown functions assembly code for linux-armhf
This allows linux loaders built for armhf architectures to support unknown
physical device and device functions. The previous fallback path could have
worked for armhf (using the appropriate compiler flags) so as to not disrupt
users, the aarch64 assembly code for unknown functions was ported to armhf.

Since CI support is limited, this was manually built & tested using on a
raspberry pi 4B.