build: Update to header 1.3.293
loader_extension_generator.py needed a tweak to handle boolean logic in
the vk.xml extension dependency attributes.  Now that the dependency is
encoded in vk.xml we don't need to manually turn on device_group for
drivers >= 1.1. We DO need to track the driver api version since the xml
expressions references it.