Define _FILE_OFFSET_BITS=64 for 32 bit linux builds
Certain file systems (such as bcachefs) use 64 bit inodes, which glibc truncates
to 32 bits by default. This causes EOVERFLOW to be returned from readdir, causing
not all files to be read from a directory.

The fix is to use the 64 bit versions of file system access functions, by defining
_FILE_OFFSET_BITS=64 in 32 bit builds.

Because this issue does not affect 64 bit builds, it is only applied to 32 bit
builds.

The test framework required explicitly referencing readdir64, the Large File
Support version of readdir found in glibc due to the function interposition that
the test framework does.

Additionally, readdir now will check errno after being called and report the
error code if it is non-zero, reporting any other issues encountered to the user
through the error stream. This doesn't apply to Windows due to implementing
readdir ontop of _findnext, which behaves differently.