Only add include paths for xcb, x11, & directfb instead of linking
Previously, the loader linked to various windowing libraries but never
made any calls into them. These can safely be removed, as including
the libraries locations in the header search paths is sufficient.

Because vulkan_core.h includes directfb.h without a path, it is
necessary to add the search path found from pkg-config as well as
the same path with `/directfb` appended to get vulkan_core.h to be
able to find the header. This may be a bug in the vulkan_core.h
or just an inconsistency with the way directfb is packaged on
ubuntu 24, the system the change was developed on.