Emulate VK_EXT_surface_maintenance1 if not supported by a driver
The VK_EXT_surface_maintance1 extension is an instance level extension,
which means the loader will advertise support for it when at least one
driver supports it. However, there is presently no way for an application
to know which VkPhysicalDevice's will fill out the structs the extension
defines and which VkPhysicalDevice's will ignore the structs.

Add support for testing the surface_maintenance1 implementation - making
sure that a driver which supports the extension isn't affected and
correctly filling out the structs for drivers which do and do not support
vkGetPhysicalDeviceSurfaceCapabilities2KHR as well.