Add Ninja + MSVC/clang-cl workaround for MARMASM
The workaround that enables ninja + MSVC/clang-cl support in the loader
only applied to MASM, which isn't helpful for MARMASM builds.

This workaround did not get added to the build as ninja support was not
considered required for release of the ARM SDK. However, other parties
have started using the ARM build of the loader and reported the issue,
necessitating this fix.