Refactor instance extension checks into single function
Loader now generates a struct that contains of bools for each instance
extension and a function to set the bool to true if the extension is in
the list passed in. This consolidates that logic into one place rather
than having it spread out across the codebase. It also enable further
changes that will require each drivers list of supported instance
extensions.