Track the extension used to create each VkSurfaceKHR
The Vulkan-Loader now doesn't call into drivers if they do not support the
extension that was used to create the VkSurfaceKHR handle. This prevents
crashes from occuring where a driver is called using a surface that it does
not know about, due to the driver not supporting the surface extension.

Because the specification requires that VkSurfaceKHR must be a valid handle,
the loader should not allow calls down into drivers which cannot know about
the surface. Instead it should "emulate" the call as appropriate, returning
0 for counts and setting to 0 any structures that were passed in.