Fix VK_INSTANCE_LAYERS bug when using vk_loader_settings.json
The issue was that copy of the VK_INSTANCE_LAYERS would be 'iterated' to
find the next path, but wouldn't be reset back to the beginning of the
stack allocation when going to the next iteration of the loop. By
creating a separate char* to hold the "current path" pointer, the start
of the allocation isn't lost.

Change-Id: Ic3f0b818bda2f0a017bc30cb1f315b245008e9f7