Make vkEnumInstExtProps return VK_INCOMPLETE correctly
The logic for whether to return VK_INCOMPLETE is straightforward - if not
enough space was available to write out all available instance layers,
write what there is space for and return VK_INCOMPLETE. The previous logic
did not take into account the way the loader settings file affects which
layers are available. The revised logic now correctly handles that situation.