Re-add ReadDir interception on macOS
This was disabled due to issues with address sanitizer, which appear to no longer
exist. The effects of intercepting readdir are an important part of the test
framework shim, so disabling that makes some tests unreliable. It would be better
to disable address sanitizer in CI where it is found to be an issue rather than
disable readdir interception.