Use VK_LOADER_DEBUG if settings stderr_log is empty
If a loader settings file was present, the logging settings of the file
were used irregardless of whether the settings file actually specified.

The loader will now check to see if the settings stderr_log value is at
least non-zero before deciding to skip a message, allowing for the
VK_LOADER_DEBUG filter to be applied in case if it is zero.