Fix CMake 3.31 builds for WIN32
A recent fix to CMake for ASM static libraries on Windows broke
the existing build.

Easiest solution I found is to remove the usage of ASM static
libraries from the loader.

This is already what the UNIX build does.

This required using generator expressions to NOT pass along
compiler options that aren't needed for assembly files.