Log Settings file Stderr Log Filters
This commit also pulls out the logic that loader_log uses to create the
ERROR | WARNING | <etc> string so that it can be reused in the logging
of the settings file stderr filters. To accomplish this, the logic was
changed slightly to assume that all debug masks are possible. This also
means adding an assert to make sure loader_log isn't called with
multiple flags set, just as a heads up that it isn't desireable.

The logging only occurs if any filters are set.

The settings file also now includes the type of layer in the log, either
explicit or implicit (based solely on the loader settings file value of
treat_as_implicit_layer).