Make Settings file "ON" take priority over disable env-var
The VK_LOADER_LAYERS_DISABLE env-var was disabling layers even if the
settings file was turning them on. The fix is a simple change in logic
to only check for the layer disable if the layer wasn't set to "ON".
Checking for "OFF" isn't needed since that happens above this check.