Correctly check if settings file exists on windows
While there was logic to check if the vk_loader_settings.json file existed,
it was not being run at an appropriate time and with another expression that
caused the check to never happen.

This commit also logs that the loader fails to find the loader settings file, and
if it finds a registry entry, if the file doesn't exist.