Log when Filter Env-Vars act when settings file is active
The logging to indicate that VK_LOADER_LAYERS_ENABLE and VK_LOADER_LAYERS_DISABLE
turned on/off layers was not occuring when the settings file was active. This
is due to layer enablement being done in `enable_correct_layers_from_settings()`,
a newer function. which lacked the log functions.
This commit rectifies the lack of logging in the new function.