Modify cJSON to not surround strings with ""
This is a modification the loader does to reduce the number of string
copies that are required. Since the loader doesn't care about the quotes,
they are simply omitted, making usage of the strings much easier.

Note that this change was done to the previously checked in version of
cJSON, so these changes are the same conceptual changes made to the newer
version of the cJSON library.