Fix O(n^2) json parsing by not using cJSON_GetArrayItem
The implementation of cJSON_GetArrayItem is very simple - it finds the
node at an index by iterating the linked-list of child nodes index
number of times. This is problematic because it means using
cJSON_GetArrayItem(node, i) to iterate all elements becomes a O(n^2).
The fix is to use the helper function cJSON_ArrayForEach which directly
iterates on the linked list.

This change takes time to refactor the get_loader_settings() logic that
handles whether the file contains just the settings object or an array
of settings objects, making the logic easier to follow.