Detect when recursive meta-layer's are present
If meta-layer A has meta-layer B in its component layer list, and
meta-layer B has meta-layer A in its component layer list, this would
cause a stack overflow in verify_meta_layer_component_layers() since
it would recursively check the meta layer of the other layer.

Fixing it means checking that the a component layer is a meta layer,
and if so, checking the component layer's component layers for the meta
layer being verified. Granted, this check is only 1 deep so more
complex recursive checks may be needed.