Make Xrandr not implicitly required when x11 is used
The previous logic made xrandr a required header whenever x11 was
available which doesn't honestly convey build requirements to users.

This commit add BUILD_WSI_XLIB_XRANDR_SUPPORT as a build option
that defaults to ON (for platforms which support x11) so the xrandr
requirement can be configured by users directly.