Use 32 bit pkg-config in linux 32 bit CI jobs
PKG_CONFIG_EXECTUABLE is a CMake variable that sets the pkg-config
executable used, allowing the use of the 32 bit pkg-config when looking
for pkg-config packages. The prior version of the 32 bit job actually
used the 64 bit packages but that happened to work okay because the only
things used from the 64 bit packages were header files, which are arch
agnostic.