Allow tests to directly call loader internal functions
This allows the fuzz tests to faithfully recreate the calling environment
of OSS-Fuzz, meaning reproduction of failing fuzz tests can be done in the
repo for regression testing.

The implementation of this requires exporting several functions that the
fuzz tests need which is done with a new macro that is only active when
a specific compiler definition is set, which the build does when testing
is active.