Support VK_GOOGLE_surfaceless_query
Requires bypassing the logic which dereferences VkSurfaceKHR in
vkGetPhysicalDeviceSurfaceFormatsKHR and
vkGetPhysicalDeviceSurfacePresentModesKHR. This should have been done
closer to the release of VK_GOOGLE_surfaceless_query, but because that
extension was meant for Android, which isn't supported by this loader,
it was not done. That said, SwiftShader can support the extension, so
support is now being added in this commit.