Make emulate_VK_EXT_surface_maintenance1 more comply to vk spec
The emulation function does not comply with specifications.
The spec asks ```The implementation must include the present mode
passed to VkSurfacePresentModeEXT in pPresentModes,
unless presentModeCount is zero.``` But now the implementation here
will always include the present mode regardless of whether
presentModeCount is zero.
According to spec ```If the value of presentModeCount is less than
the number of compatible present modes that are supported, at most
presentModeCount values will be written to pPresentModes.``` When
presentModeCount is 0, at most 0 values will be written to
pPresentModes, in other words the implementation should do nothing
in this situation.
This commit fixed the above issue.