Refactor codegen scripts to use vulkan_object.py
Requires hardcoding a few things to make it work, most notably:
* Missing Extension Requires Depends information
* Missing aliases for handles

On the whole, the codegen outputs *almost* identical files from before,
making the refactor not able to introduce issues into the output. The
changes that do occur to remove duplicated if statements from aliased
handles or for comments and whitespace.