Fix pAllocator in lazily allocated ICD surfaces
The pAllocator passed into each surface creation function would be a
pointer to the instance's pAllocator callbacks, regardless of whether
they were valid.

This commit saves the VkAllocationCallbacks passed into the surface
create function so that it can be used when the surface handle is
actually created in the driver.