Add device and drivers fields to settings file
The loader settings file now can specify additional drivers to include, whether
to use those drivers exclusively, and filter/sort the VkPhysicalDevices using
the settings file provided list of deviceUUID's.

This lets the settings file control which drivers and/or devices are reported
by the loader, giving greater control over the runtime behavior of the loader.

device_configurations is an array of JSON objects which contain two fields
each, deviceUUID and deviceName.
deviceUUID is an array of 16 uint8_t's (expressed simply as numbers) which
takes the value of the
VkPhysicalDeviceVulkan11Properties::deviceUUID struct member.
deviceName is a
string taken from VkPhysicalDeviceProperties::deviceName. Only deviceUUID is
used to filter and sort VkPhysicalDevices, deviceName is for convenience and
ease of understanding the JSON loader settings file.

additional_drivers is an array of JSON object which contain just one field,
path, which is a path to the Manifest ICD JSON file for a driver.

additional_drivers_use_exclusively is a boolean field indicating whether the
loader should:
if false (default if unspecified) - look for drivers in the normal locations
and append the  additional_drivers to the back of the driver list.
if true - only use drivers in the additional_drivers, ignoring all other
driver search locations.