Store ICDs in the loading order
This commit switch from prepending newly loaded ICDs to the ICD
list, to appending them.

Both options are fine since the order VkPhysicalDevices appear
is unspecified, but by using an append operation the
VkPhysicalDevices are returned in the same order as the drivers
are loaded.

See https://github.com/KhronosGroup/Vulkan-Loader/issues/1725