Cleanup command extension checking of device functions
Remove some TODO's about needing the detailed dependency info of a command.
Commands from extensions are available if any of the extensions they come
from are enabled. The extensions themselves have complex requirements for enabling,
but because the loader tries to do basic validation, it should only check that the primary
extensions are available. This is so that the loader doesn't replicate all of the complicated
dependency checking logic that already present in the Validation Layer.