Allow VK_INSTANCE_LAYERS to order implicit layers
Implicit layers previously were always loaded closest to the application regardless of
whether they appeared in any environment variables. This commit allows users to order
implicit layers using VK_INSTANCE_LAYERS, so that if a user has put an implicit layer in
that environment variable, that layer will be loaded respecting its order relative to other
layers in the env-var. If VK_INSTANCE_LAYERS contains just a single implicit layer, this
will cause that implicit layer to be loaded after all other implicit layers, and before any
app-enabled layers.

This commit also updates the documentation to reflect the un-deprecated nature of
VK_INSTANCE_LAYERS as well as other fixes related to this change.