Add driverUUID and driverVersion to settings file device configuration parameters
deviceUUID is not sufficient to be able to distinguish VkPhysicalDevices as there can be multiple
ICD's that report the same VkPhysicalDevice. This can easily occur when a development build of
a driver is present alongside a stable build on a system. driverUUID then helps differentiate ICD's
reporting the same VkPhysicalDevice, but isn't sufficient as some drivers use the same driverUUID
across all builds (RADV from Mesa as example). Thus, driverVersion is used for that purpose.