Remove duplicate driver workaround for macOS
On macOS, applications can bundle drivers with themselves. When they do
this and also install the same driver to a global location, the loader
will load 2 separate drivers which happen to be the same binary,
causing the linker to emit warnings.

The fix was to simply only load one driver if the app was bundled.
Because there was only one macOS driver available, this worked. Now that
both KosmicKrisp and LavaPipe are available, this workaround is no
longer appropriate. If apps bundle multiple drivers, they would be
unable to use more than 1.