Fix devices in the settings file removing layers
The settings file supports multiple different kind of settings, layers,
devices, and drivers, mixing and matching each kind. This didn't work
when there was only a single boolean to determine if the settings are
active. Adding a separate boolean for layers and devices in the settings
allows for only running settings file logic when the relevant parts of
the settings file are present, so that a settings file with only devices
doesn't interfere with the layer enumeration and vice versa.
Additionally, using booleans instead of just checking if the list of
layer/device settings are non-empty allows the user to specify
"no layers" and "no devices" on the system.