Fix duplicate layer output with the settings file
When there exists a layer in the settings file with the same name as a
layer in the default search paths with a different library_path, the
loader was mistaking it as a 'distinct' layer. Because the settings file
allows specifying layers by their path & whether to enable/disable them,
it is possible to exactly specify which layers will load, including
duplicates.