Cleanup test layer creation
Move the ManifestLayer out of the TestLayerDetails, as well as drop
TestLayerDetails in favor of BinaryCreateArgs, since its the same as for
test ICD creation. This makes configuring test layer creation simpler
since everything isn't bound to one type.

Also dropped the json name from most layers in the tests because that
isn't an active part of the test. Simplifies test creation to not need
to specify a json file name when it doesn't matter for the purpose of
the test. Tests which do care (eg, ones which inspect the log) still
can set a name and it is respected. Otherwise, layers will use
"test_layer_x.json" as the default name, where x is a number used to
distinguish created layers.